requirements.txt
chromadb==0.5.3
googletrans==3.0.0
langchain==0.3.7
langchain_community==0.3.7
numpy==1.23.5
pandas==1.5.3
pymongo==4.3.3
PyPDF2==3.0.1
python_docx==1.1.2
streamlit==1.39.0
streamlit_option_menu==0.3.13

